Hardened tests
