+++ .changeset/1237-flaky-tests.md
@@ -1,0 +1,6 @@
+---
+"@synergy-design-system/components": none
+"@synergy-design-system/mcp": none
+---
+
+chore: 🔧 flaky tests (#1237)
