+++ packages/components/package.json
@@ -45,7 +45,7 @@
     "figma-export:icons-2025-outline": "pnpm exec figma connect publish -c ./code-connect/figma.config.icons.2025-outline.json --batch-size 100",
     "report-bundle-size": "pnpm exec size-limit --json",
     "build": "node scripts/build.js",
-    "test": "web-test-runner --group default --config ./web-test-runner.synergy.config.js",
+    "test": "web-test-runner --group default --config ./web-test-runner.config.js",
     "test.watch": "pnpm test --watch",
     "lint:css": "stylelint \"src/**/*.{css}\" \"src/**/*.styles.ts\"",
     "lint:html": "html-validate -c ./code-quality/html-validate/.htmlvalidate.js ./demos",
