+++ packages/mcp/metadata/checksum.txt
@@ -1,1 +1,1 @@
-d1dbcb89316700587875bfb327172531
+c365690db814db34e8c4d548355f21f9
